chore: move typography to root of src (#2537)
