+++ packages/react/src/components/ErrorSummary/ErrorSummaryHeading.tsx
@@ -1,5 +1,5 @@
 import { forwardRef, useContext, useEffect } from 'react';
-import { Heading, type HeadingProps } from '../Typography/Heading';
+import { Heading, type HeadingProps } from '../Heading';
 import { ErrorSummaryContext, type ErrorSummaryProps } from './ErrorSummary';
 
 export type ErrorSummaryHeadingProps = HeadingProps;
