+++ packages/react/src/components/List/List.test.tsx
@@ -3,7 +3,6 @@ import { render as renderRtl, screen } from '@testing-library/react';
 import type { ListUnorderedProps } from './Lists';
 
 import { List } from '.';
-import { Heading } from '../Typography/Heading';
 
 const render = (props: Partial<ListUnorderedProps> = {}) => {
   const allProps: ListUnorderedProps = {
