--- packages/react/src/components/Typography/index.ts
@@ -1,4 +1,0 @@
-export * from './Paragraph';
-export * from './Heading';
-export * from './Label';
-export * from './ValidationMessage';
