+++ packages/react/src/components/form/Combobox/internal/ComboboxLabel.tsx
@@ -1,7 +1,8 @@
 import { PadlockLockedFillIcon } from '@navikt/aksel-icons';
 import cl from 'clsx/lite';
 
-import { Label, Paragraph } from '../../../Typography';
+import { Label } from '../../../Label';
+import { Paragraph } from '../../../Paragraph';
 import type { useFormField } from '../../useFormField';
 import type { ComboboxProps } from '../Combobox';
