+++ packages/react/src/components/index.ts
@@ -7,7 +7,10 @@ export * from './loaders/Spinner';
 export * from './loaders/Skeleton';
 export * from './Link';
 export * from './List';
-export * from './Typography/';
+export * from './Label';
+export * from './Heading';
+export * from './Paragraph';
+export * from './ValidationMessage';
 export * from './Accordion';
 export * from './form/Select';
 export * from './Alert';
