+++ packages/react/stories/Typography.mdx
@@ -1,10 +1,10 @@
 import { Meta, Canvas, Controls } from '@storybook/blocks';
 
 import * as typographyStories from './Typography.stories';
-import * as HeadingStories from './Heading/Heading.stories';
-import * as ParagraphStories from './Paragraph/Paragraph.stories';
-import * as LabelStories from './Label//Label.stories';
-import * as ValidationMessageStories from './ValidationMessage/ValidationMessage.stories';
+import * as HeadingStories from '../src/components/Heading/Heading.stories';
+import * as ParagraphStories from '../src/components/Paragraph/Paragraph.stories';
+import * as LabelStories from '../src/components/Label/Label.stories';
+import * as ValidationMessageStories from '../src/components/ValidationMessage/ValidationMessage.stories';
 
 <Meta of={typographyStories} />
